Conclusion is now editable in restriction cases
