+++ libs/judicial-system/formatters/src/lib/formatters.spec.ts
@@ -14,6 +14,7 @@ import {
   formatCustodyRestrictions,
   formatAlternativeTravelBanRestrictions,
   formatProsecutorDemands,
+  formatConclusion,
 } from './formatters'
 
 describe('formatDate', () => {
@@ -556,5 +557,404 @@ describe('formatGender', () => {
 
     // Assert
     expect(r).toBe('Kynsegin/Annað')
+  })
+})
+
+describe('formatConclusion', () => {
+  test('should format conclusion for a rejected case', () => {
+    // Arrange
+    const type = CaseType.CUSTODY
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.REJECTING
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = false
+    const isExtension = false
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      undefined,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kröfu um að kærði, Glanni Glæpur, kt. 010101-0000, sæti gæsluvarðhaldi er hafnað.',
+    )
+  })
+
+  test('should format conclusion for an accepted case without isolation', () => {
+    // Arrange
+    const type = CaseType.CUSTODY
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.ACCEPTING
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = false
+    const isExtension = false
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      undefined,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kærði, Glanni Glæpur, kt. 010101-0000, skal sæta gæsluvarðhaldi, þó ekki lengur en til þriðjudagsins 22. desember 2020, kl. 11:23.',
+    )
+  })
+
+  test('should format conclusion for an accepted case with isolation', () => {
+    // Arrange
+    const type = CaseType.CUSTODY
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.ACCEPTING
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = true
+    const isExtension = false
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      undefined,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kærði, Glanni Glæpur, kt. 010101-0000, skal sæta gæsluvarðhaldi, þó ekki lengur en til þriðjudagsins 22. desember 2020, kl. 11:23. Kærði skal sæta einangrun á meðan á gæsluvarðhaldinu stendur.',
+    )
+  })
+
+  test('should format conclusion for an accepted case with isolation and the isolation ends before the custody does', () => {
+    // Arrange
+    const type = CaseType.CUSTODY
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.ACCEPTING
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = true
+    const isExtension = false
+    const isolationToDate = new Date('2020-12-20T15:39')
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      undefined,
+      isolationToDate,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kærði, Glanni Glæpur, kt. 010101-0000, skal sæta gæsluvarðhaldi, þó ekki lengur en til þriðjudagsins 22. desember 2020, kl. 11:23. Kærði skal sæta einangrun ekki lengur en til sunnudagsins 20. desember 2020, kl. 15:39.',
+    )
+  })
+
+  test('should format conclusion for a case where custody is rejected, but alternative travel ban accepted', () => {
+    // Arrange
+    const type = CaseType.CUSTODY
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.ACCEPTING_ALTERNATIVE_TRAVEL_BAN
+    const validToDate = new Date('2021-01-29T13:03')
+    const isolation = false
+    const isExtension = false
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      undefined,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kærði, Glanni Glæpur, kt. 010101-0000, skal sæta farbanni, þó ekki lengur en til föstudagsins 29. janúar 2021, kl. 13:03.',
+    )
+  })
+
+  test('should format conclusion for rejected extension', () => {
+    // Arrange
+    const type = CaseType.CUSTODY
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.REJECTING
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = false
+    const isExtension = true
+    const previousDecision = CaseDecision.ACCEPTING
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      previousDecision,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kröfu um að kærði, Glanni Glæpur, kt. 010101-0000, sæti áframhaldandi gæsluvarðhaldi er hafnað.',
+    )
+  })
+
+  test('should format conclusion for rejected extension when previous ruling was travel ban', () => {
+    // Arrange
+    const type = CaseType.CUSTODY
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.REJECTING
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = false
+    const isExtension = true
+    const previousDecision = CaseDecision.ACCEPTING_ALTERNATIVE_TRAVEL_BAN
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      previousDecision,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kröfu um að kærði, Glanni Glæpur, kt. 010101-0000, sæti gæsluvarðhaldi er hafnað.',
+    )
+  })
+
+  test('should format conclusion for accepted extension', () => {
+    // Arrange
+    const type = CaseType.CUSTODY
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.ACCEPTING
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = false
+    const isExtension = true
+    const previousDecision = CaseDecision.ACCEPTING
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      previousDecision,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kærði, Glanni Glæpur, kt. 010101-0000, skal sæta áframhaldandi gæsluvarðhaldi, þó ekki lengur en til þriðjudagsins 22. desember 2020, kl. 11:23.',
+    )
+  })
+
+  test('should format conclusion for accepted extension when previous ruling was travel ban', () => {
+    // Arrange
+    const type = CaseType.CUSTODY
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.ACCEPTING
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = false
+    const isExtension = true
+    const previousDecision = CaseDecision.ACCEPTING_ALTERNATIVE_TRAVEL_BAN
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      previousDecision,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kærði, Glanni Glæpur, kt. 010101-0000, skal sæta gæsluvarðhaldi, þó ekki lengur en til þriðjudagsins 22. desember 2020, kl. 11:23.',
+    )
+  })
+
+  test('should format conclusion for rejected extension when alternative travel ban accepted', () => {
+    // Arrange
+    const type = CaseType.CUSTODY
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.ACCEPTING_ALTERNATIVE_TRAVEL_BAN
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = false
+    const isExtension = true
+    const previousDecision = CaseDecision.ACCEPTING
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      previousDecision,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kærði, Glanni Glæpur, kt. 010101-0000, skal sæta farbanni, þó ekki lengur en til þriðjudagsins 22. desember 2020, kl. 11:23.',
+    )
+  })
+
+  test('should format conclusion for rejected extension when alternative travel ban accepted and previous ruling was travel ban', () => {
+    // Arrange
+    const type = CaseType.CUSTODY
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.ACCEPTING_ALTERNATIVE_TRAVEL_BAN
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = false
+    const isExtension = true
+    const previousDecision = CaseDecision.ACCEPTING_ALTERNATIVE_TRAVEL_BAN
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      previousDecision,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kærði, Glanni Glæpur, kt. 010101-0000, skal sæta áframhaldandi farbanni, þó ekki lengur en til þriðjudagsins 22. desember 2020, kl. 11:23.',
+    )
+  })
+  test('should format conclusion for a rejected travel ban', () => {
+    // Arrange
+    const type = CaseType.TRAVEL_BAN
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.REJECTING
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = false
+    const isExtension = false
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      undefined,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kröfu um að kærði, Glanni Glæpur, kt. 010101-0000, sæti farbanni er hafnað.',
+    )
+  })
+
+  test('should format conclusion for an accepted case without isolation', () => {
+    // Arrange
+    const type = CaseType.TRAVEL_BAN
+    const accusedNationalId = '0101010000'
+    const accusedName = 'Glanni Glæpur'
+    const accusedGender = CaseGender.MALE
+    const decision = CaseDecision.ACCEPTING
+    const validToDate = new Date('2020-12-22T11:23')
+    const isolation = false
+    const isExtension = false
+
+    // Act
+    const res = formatConclusion(
+      type,
+      accusedNationalId,
+      accusedName,
+      accusedGender,
+      decision,
+      validToDate,
+      isolation,
+      isExtension,
+      undefined,
+    )
+
+    // Assert
+    expect(res).toBe(
+      'Kærði, Glanni Glæpur, kt. 010101-0000, skal sæta farbanni, þó ekki lengur en til þriðjudagsins 22. desember 2020, kl. 11:23.',
+    )
   })
 })
